Reorganize organic jobs code
