+++ validator/app/src/compute_horde_validator/validator/tasks.py
@@ -33,7 +33,6 @@ from django.utils.timezone import now
 from compute_horde_validator.celery import app
 from compute_horde_validator.validator.locks import Locked, LockType, get_advisory_lock
 from compute_horde_validator.validator.metagraph_client import get_miner_axon_info
-from compute_horde_validator.validator.miner_client import MinerClient
 from compute_horde_validator.validator.models import (
     Cycle,
     JobFinishedReceipt,
@@ -43,6 +42,8 @@ from compute_horde_validator.validator.models import (
     SystemEvent,
     Weights,
 )
+from compute_horde_validator.validator.organic_jobs.miner_client import MinerClient
+from compute_horde_validator.validator.organic_jobs.miner_driver import execute_organic_job
 from compute_horde_validator.validator.synthetic_jobs.batch_run import (
     SYNTHETIC_JOBS_HARD_LIMIT,
     SYNTHETIC_JOBS_SOFT_LIMIT,
@@ -51,7 +52,6 @@ from compute_horde_validator.validator.synthetic_jobs.utils import (
     create_and_run_synthetic_job_batch,
 )
 
-from .miner_driver import execute_organic_job
 from .models import AdminJobRequest
 
 logger = get_task_logger(__name__)
