+++ validator/app/src/compute_horde_validator/validator/tests/helpers.py
@@ -18,12 +18,12 @@ from compute_horde.mv_protocol.validator_requests import BaseValidatorRequest
 from django.conf import settings
 from substrateinterface.exceptions import SubstrateRequestException
 
-from compute_horde_validator.validator.facilitator_api import (
+from compute_horde_validator.validator.models import SystemEvent
+from compute_horde_validator.validator.organic_jobs.facilitator_api import (
     V0FacilitatorJobRequest,
     V1FacilitatorJobRequest,
 )
-from compute_horde_validator.validator.miner_client import JobState, MinerClient
-from compute_horde_validator.validator.models import SystemEvent
+from compute_horde_validator.validator.organic_jobs.miner_client import JobState, MinerClient
 from compute_horde_validator.validator.synthetic_jobs import batch_run
 
 NUM_NEURONS = 5
